Phone URL changes

+++ Entities/Base.cs
@@ -9,7 +9,7 @@ namespace Entities
     public class Base
     {
         [Key]
-        public uint ID { get; set; }
+        public long ID { get; set; } // Long needed here because EF can't use uint 
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
         public DateTime DeletedAt { get; set; }
