+++ Entities/Models/City.cs
@@ -7,7 +7,7 @@ namespace Entities.Models
     public class City : Base
     {
         public string Name { get; set; }
-        public uint StateID { get; set; }
+        public long StateID { get; set; }
         public virtual State State { get; set; }
 
         public virtual List<Company> Companies { get; set; }
